+++ app/src/main/java/com/projeto/integrador/CadastroActivity.java
@@ -3,11 +3,13 @@ package com.projeto.integrador;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
-public class CadastroActivity extends AppCompatActivity {
+public class CadastroActivity extends DrawerCreator {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
+        if (savedInstanceState == null)
+            savedInstanceState = new Bundle();
+        savedInstanceState.putInt("layout", R.layout.activity_cadastro);
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_cadastro);
     }
 }
