--- app/src/main/java/com/projeto/integrador/Layouts/DrawerCreator.java
@@ -1,24 +1,0 @@
-package com.projeto.integrador.Layouts;
-
-import android.app.Activity;
-import android.support.annotation.NonNull;
-import android.support.design.widget.NavigationView;
-import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.ActionBarDrawerToggle;
-import android.view.LayoutInflater;
-import android.view.MenuItem;
-
-public class DrawerCreator implements NavigationView.OnNavigationItemSelectedListener{
-
-    private DrawerLayout drawerLayout;
-    private ActionBarDrawerToggle actionBarDrawerToggle;
-    private NavigationView navigationView;
-
-    public void CriandoUmDrawer(Activity activity){
-    }
-
-    @Override
-    public boolean onNavigationItemSelected(@NonNull MenuItem menuItem) {
-        return false;
-    }
-}
