+++ app/src/main/res/layout/activity_main.xml
@@ -7,19 +7,16 @@
     android:layout_height="match_parent"
     tools:context=".MainActivity">
 
-    <TextView
+    <LinearLayout
+        android:id="@+id/inclusao"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Hello World!"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        android:orientation="vertical" />
 
     <android.support.design.widget.NavigationView
+        android:id="@+id/navigationMain"
         android:layout_width="wrap_content"
         android:layout_height="match_parent"
-        android:id="@+id/navigationMain"
         android:layout_gravity="start"
         app:headerLayout="@layout/cabecalho_gaveta"
         app:menu="@menu/menu_navegacao_gaveta">
