+++ app/src/main/res/menu/menu_navegacao_gaveta.xml
@@ -13,12 +13,7 @@
 
     <item
         android:id="@+id/servicos_em_andamento"
-        android:icon="@drawable/briefcase"
-        android:title="Serviços"/>
-
-    <item
-        android:id="@+id/config"
-        android:icon="@drawable/settings"
-        android:title="Configurações" />
+        android:icon="@drawable/icon_money"
+        android:title="Orçamentos"/>
 
 </menu>
